More clean-up in the interpolation file.  The edge extension mode is no longer part of the InterpolationView class.  Instead, we rely on the free functions to wrap the image in an edge extension mode, and then pass the edge extended image to the Interpolation class.  This will leave room for more flexibility in the future.
Also added the first cut at the reworked Transform.h.  It's
substantially different, however the largest change is that you can
you crop() to select a region of the TransformView using a bounding
box.  The routines for computing the extent of the transformed
bounding box have been broken out into seperate functions.