Spent some time updating the build system so that VW can be built on the IRG cluster machines (specifically redhat ws 4 with gcc 3.4).
Fixed a bug in a regexp in the AX_BOOST_PKG so that the script can
find very funky boost library names with like
libboost_thread-gcc-mt-1_33_1.a

Fixed a bug in the AX_PTHREAD macro that prevented the proper linker
flags from being discovered.  The compile test was only sufficient to
verify the cppflags, and not the ldflags.  The current solution will
probably work in the general case, but we might want to think about
how to improve this test in the future.  Talk to me for more details.

Fixed a filename case bug that was masked by Darwin's non-case
sensitivity.

Added cppflags to the tools makefile so that it could be compiled.