Added new image statistics: standard deviation and sum
Added the basics for two new camera models: fisheye and linescan
  The linescan model is generic enough to be used for the HRSC imager.
  Fisheye is just for fun.  Neither are yet complete.

Small formatting change to the optimized correlater text output

Added a new "Pyramid" correlator that can be used to automatically
search for a tight disparity search range in any image.

Started work on a "Multiresolution" correlator that is sort of a
hybrid of the normal correlator and the pyramid.  This is just an
experiment, and it doesn't work yet.