Improved documentation of interest point module. docs/workbook now includes interestpoint_module.tex, a draft of the API documentation for the module. The header files are commented better. The code itself is mostly uncharged, except for small fixes (like making the write_images member functions of the Detector classes public instead of protected).
The API will probably change a bit before release, and
one thing that does need to be cleared up is the
interface to interest_points in Detector.h, which is
currently overloaded. The second version was added as
a quick compatibility fix for the stereo pipeline, but
makes the first version impossible to use.