Added logic to the Statistics to ignore transparent pixels when the pixel type has alpha.
Added a polar stereographic projection for the georeference.

MATT--> CHECK THIS CODE: Added some values to Fundamental types for
determining the absolute limits of various data types.  This is used
in the statistics functions to pick initial min/max values that are as
positive or negative as possible for various channel types.

Made some changes to the interest point code so that it compiles with
the stereo pipeline.