Checking in another major overhaul to the ThreadPool.  There is no longer a threadpool class -- it has been replaced with a simpler WorkQueue object.  There are a few useful subclasses of WorkQueue such as FifoWorkQueue and OrderedWorkQueue.
Added a first cut at a threaded version of write_image, which is currently called block_write_image().  I have yet to add different versions of this function that take filenames, bounding boxes, etc.

Added a flag to set_tiled_write() for OpenEXR that allows out-of-order writing of image blocks to disk.