Removed some unused arguments in compute_camera_bbox()
Added a feature to OrthoImageView to check for "transparent" pixels in the underlying DEM.  If transparent pixels are found, then the OrthoImageView returns a default constructed pixel for those areas.  (The intuition here is that you should not ever render orthoimage data where there is no underlying terrain information).   This feature can be used along with the new PixelMask<> code I'm about to check in.  More info to follow.