clean up includes.
this drops something like 100 unnecessary includes. I tried to be logical
when removing stuff- the idea was to flatten the tree a bit. If A->B, but
nothing directly declared in B is used, drop the include (or include
something higher in the tree). I've also attached the script I used to aid
this- it sorts the headers by how deep into the tree they are, where depth
is max(children.depth)+1. cd to the top-level source dir and run
./config/include-graph.py. The script is pretty fragile, since it mostly
relies on a grep/sed operation.

I'd bet money that there are a few missing includes. Sorry. Unit testing
will shake out the problems ;)