Added a simple Particle Swarm Optimization (PSO) - a global, gradient-free stochastic optimizer.
Includes some tests - though these don't test the internals
of the algorithm as such but rather just note that it indeed tends
to minimize functions. These tests can fail every once in a while
as PSO uses (lots of) random numbers - the results are not the same
from one run to the next.