Add methods for calibrating some or all parameters of a PinholeModel based on known points in 3D space and the corresponding pixel locations.
This includes the required code for (de)serializing parameters of a
PinholeModel into a vw::Vector and the methods and functors needed
for applyingthe Levenberg Marquardt optimizer to the problem and a
variant using RANSAC for outlier detection.

Also included are some tests for both the serialization and the
calibration methods.