disable strict aliasing by default
We are violating strict aliasing rules somewhere. The g++ 4.4.1 (on
RHEL5), with -O3 (or, really, -O1 -finline-functions -ftree-pre
-fstrict-aliasing) generates an infinite loop in one of the inlined copies
of dot_prod called indirectly from TestStereoModel::test_pinhole_stereo_model,
inside StereoModel::operator()(Vector2 const&, Vector2 const&, double&) const;

We need -finline-functions to get rid of all the accessors, and
-fno-tree-pre will paper over the problem without fixing it.
-fno-strict-aliasing will prevent gcc from making dangerous assumptions
about the code.

To be clear, though, this is still a stopgap measure. We'll pay a speed
penalty for this, which is why I've left it easily re-enableable

I've been playing with gcc 4.4, and it detects aliases that earlier gccs
don't (http://gcc.gnu.org/gcc-4.4/porting_to.html). I have to check them
for false positives, but hopefully it'll lead to solving it.