fix GeoTransform for UTM zones far from poles
For utm proj.4 will return an error if you try to project lonlat->utm wher
the lonlat is more than 90 degrees away from the utm meridian.

This pole check should be generalized into a discontinuity check for any
projection. I filed a bug so we don't lose it.

This also adds a test that would fail without this fix.