I changed the way sparse systems are solved. Now, rather than doing the full solution in
sparse_solve, 

there is a helper routine called 

sparse_solve_ldl

which performs backwards and forwards substitution 
and assumes it is given the ldl decomposition. 

The sparse_solve routine now does the 
ldl_decomposition, and then calls 
sparse_solve_ldl. 

In addition, I implemented a 
multi_sparse_solve routine, which 
can sparsely solve systems of the form 
AX = B, 

where A is sparse but B is not 
(In particular, if B is the identity,
the routine will find the non-sparse
inverse for A)

This routine is modeled after 
multi_solve_symmetric 
located in LinearAlgebra.h. 

I did not write unit tests in the testing directory. 

I have tested the code against regular inversion, and the results
match exactly.