Providing an alternative CameraBBox that accepts a DEM and can perform intersection against it.
CameraBBox provides a bounding box on the map where a camera's image intersects. Map topography plays a big part in this bounding box when the camera is close to the ground (ie LCROSS). I've added a LMA algorithm that works out the correct intersection. It appears to be working but the LMA has a potental of not finding exactly the correct result if the seed solution is bad (seed solution being old CameraBBox's method).