timed_wait never worked. Fixed, and added a test.
It incorrectly called wait(). In fixing this, I also had to fix the bug
the TODOs reference. The problem was that scoped_lock was actually a
typedef inside boost::mutex, and everything needed to know the real class
of the mutex.  In addition, boost::condition needed the explicit lock()
and unlock() on the mutex and lock.