Added new Scaled Gradient Descriptor to Interest Point Module.
This is a lazy person's gradient descriptor. I seems to me to be a mashup of GLOH and Daisy. It still needs tuning up and no way matches state of the art interest point methods. Yet this produces twice as many matches than our patch descriptor. SGrad detector also can be matched a heck of a lot faster than patch because this is 180 elements versus patch's 1600.

IP module would best be improved with a new detector next (OBALoG).