Added Timers in Sparse BundleAdjustment. Modified SparseSkylineMatrix into MatrixSparseSkyline.
Modification to MatrixSparseSkyline brings the skyline to be built on top of MatrixBase. I've switched from compressed_matrix to generalized vector of compressed vectors as insertation time is supposed to be faster. SparseSkyline operations now have the option to accept outside skyline vectors. SparseSkyline being built on top of MatrixBase and the use of outside skyline vector is for the upcomming ReorderMatrix and ReorderVector wrapper.