Added test for error that will crop up with trying to access const iterator through select_col() of an object with a funky iterator (like MatrixTranspose). I've also added a controversial fix for this test.
The problem is when const begin is accessed through select_col. It can't provide a const iterator to it's inner type as it doesn't own a const reference to the inner type. I've added a const_cast that adds a const to the inner type when accessing const begin/end. I know I'll get lacerated if I cast away const, but what if I add it?