Move timeout configuration from endpoint to client
Basically, this makes CallMethod obey a configured timeout. This is
necessary to prevent apache from hanging past the client timeout.

The defaults are set to what CallMethod is currently using, with one
caveat. Sequence failures now count toward the retry count. This is
necessary because I need to ensure an upper bound on delay (tries *
timeout).

This might mean that the default number of tries needs to be increased- I
don't know how often non-fatal sequence failures happen in practice.