Make sure lonlat_to_point doesn't return bad values
It could return a completely invalid (-1,-1) value with no warning or
anything. I also changed when the clamp happens, since we should maintain
precision as long as possible; it's even possible that this clamp order
discrepancy triggered the problem the workaround was supposed to solve.
Don't know for sure, though.