Added Pixel comparison helpers
It's a little ugly in that the inner function can't be passed. There are specific equal and near comparison operators so that the comparitor can be templatized on the PixelType.

Also .. funny thing .. expect eq doesn't report meaningful errors when the inner type is char. Expect Pixel Eq has a specific implementation to fix that.