BundleAdjustment moved to own module
Yikes! BA code makes a lot of dependencies on InterestPoint / Camera / Cartography / and Stereo. In order to have some sanity to the build Bundle Adjustment was isolated so that it's just known that it's a very high level library. This allows Camera Module to be a mid level ish library.

Some other renames also happened to avoid typing bundle adjustment twice.

Bundle Adjustment module uses the ba namespace.