CDFAccum support arbituary number of quantiles
More quantiles provides more potential for the CDF to be accurate (needs more samples though). I also changed the accumulators tests so that the random number generator doesn't get seeded with time. This should make it a predictable random .. so that our tests past with certainty and not just with great likelihood.