Fixed allocation size
Cache object now gets the proper size of the image that was
registered. I also shrank that hard coded 512MB of texture memory to
256 MB. This is a non ideal solution, but finding out the amount of
memory on a GPU is vendor specific. OpenGL doesn't want to tell the
users how much RAM they have to work with as it will automatically
throw it back on to system memory. 256 MB seems to be a sweet spot and
vwv no longer crashes after extended use on bespin. Virtual Memory tops out at 1.4Gb.