BA Reporter switched to CDF Accumulator
BA now reports approximate mean and standard deviation. It also prints out the [0, 0.25, 0.5, 0.75, 1.0] quantiles of the CDF. This seems more descriptive in my opinion. Also this means only a sigle pass through of the measurements is required.

A better next step would be to have the models setup a iterator facade so that a std::vector of measurements is not required.