Added remove descriptor functor
Used this functor to clear up some memory on the control network loader. Control Network loader gets 99% of the way towards building a CNET for A15 and A16 before it crashes. 2.2 GB is used to filter the measurements. The problem is flipping the camera relations network into a control network in one go. It needs to be done incremently. Maybe writing of the CNET can be done incrementally as well.