Finish orthoproject tests
It's a little weak .. it's hard to quantify an interpolation error. There is one pixel test to check for an interpolation error. There's also one pixel test to check triangulation.

There's a series of traits check that make sure edge extension and interpolation view are inline with correct operations as well.