in-class static initialization is always a bad idea
Even with integral types, in-class static init doesn't work right. It only
works if every use of the constant is inlined. If you ever use that
constant from a place where you need an address (ie, reference or pointer)
it will fail to link with an undefined reference. Values may become
references in surprising places if optimization is turned on.

Take-away: don't ever use static in-class initialization. At all. (lame).