Added write pool
Write pool is the number of slots for threads waiting write. This allows more control on the amount of memory being used. Most importantly it makes it easier to keep the computer tasked .. but that's hard to explain.

Example:
Computer with 8 cores but only want to keep it tasked at 600% with VW. Previously in VW you would just set the number of threads to 6 and when one task took much longer the VW process would throttle down. Now you can set number of threads to 6 and writing pool to 20, so that when one process takes a really long time, VW can still keep working on other blocks until the write pool fills up.