Have SOAD return CorrelatorAccumlator type
This keeps us from returning doubles all of the time. Also the
Correlator Failure value has been removed and replaced with an
overloaded CorrelatorFailureValue. Meaning if we're working with a FP
value we return a quiet NaN when the correlator fails. With INT
correlator we return max int.

Also there's a bug fix in here to avoid a SIGFPE that happens in
parabola subpixel correlator.