Added set_nodata_value options to DiskImageResourceGDAL
We had the read .. but never the write. Also added TestGDALFeatures
that checks that the whole nodata loop works.

I'm not pleased with this implementation. I think that the nodata
functions should be templated on a pixel type and be able to convert
the nodata correctly to which ever pixel type the user provides.