Replace transaction typedefs with concrete type
Before, uint32 was used for concrete transactions and int32 included the
ability to pass -1. Those use-cases are replaced by Transaction and
TransactionOrNeg, respectively. Transactions are convertable to uint32,
TransactionOrNeg requires that you check for -1 and then promote it to a
Transaction (or just promote it, and let it throw if it's -1).