clean up blob size handling
This should fix the end-of-blob corruption issue.  Blobs are not
restricted to being less than numeric_limits<int32>::max(), and some parts
of the size handling held it as an int32. There's a pretty good chance
that any blob bigger than that is damaged near the end in some way.

I also backed off the default blob size limit to 1.75GB, because we can't
actually prevent clients from making the blobs bigger than 2GB (since we
currently give them full control of the blob).