Added plate2kml
This added the dependency of LibKML. If not found, plate2kml is not
built.

This is a bare bones working executable that converts Plate files to
KML trees like those created by image2qtree. There such much more that
could be done with LibKML. Plate2kml could become a webservice if we
wanted. There's no reason to copy all of image data into a new tree.

At the current moment plate2kml pulls the top transaction from the
plate. It is also single threaded. I'm not sure if LibKML is
threadsafe but it could be future development.