Added simpler test to detect bug in GCC4.0.1
The bug happens only on GCC4.0.1 with -01 or higher. Replacing -01
with the individual optimization flags causes the problem to
disappear. I've been unable to debug past that point. For now, there
is a warning in the unit tests that recommends the user to upgrade GCC
(which is available via XCode).