Enforce correct const use inside math module.
A const method will always call a const method inside a child
object. So Vectors and Matrices holding references to a child type now
have child() methods that provide a const and non const interface.

This fixes potential bugs like normal const issues, but also cases
where the const reference and reference are not the same object. An
example of that would be a boost::compressed_vector.