Make the TypeDeduction system 32/64bit safe
It was failing the test I wrote in TestVector.cxx because on my system
'signed int' and 'signed int long' are the same thing. This meant that
the type deduction system had no support for 64 int types! So double
vector arithmetic against int64 types was returning int64 because our
type deduction system thought int64 was a custom user type.

Oi!