Add FLANN as an optional dependency
When FLANN is available, it over-rides the use of KD-Trees in the
interest point matcher.

Here are the timings for 2 sets of 2800 interest points with 128
element descriptors:

KD-Tree:     5.1 seconds
Brute Force: 2.5 seconds
FLANN Tree:  0.2 seconds

FLANN is an incredible improvement over KD-Tree mostly just because
KD-Trees really suck with high dimensionality data. That case is best
for approximated nearest neighbors.