Revert "Added a safety in DIRGDAL to avoid excessive tile sizes"
This reverts commit 381c6861d34baab1a56af19580655a1ffd0bab9f.

It turns out this is a bad idea because of the Pyramid Correlator. The
default tile size that GDAL chooses is a single line. That just
doesn't work at all for the Pyramid Correlator.

It seems this is a case where it is better to let the user fail if
they requests a tile size bigger than the input.