Change GeoReference bounding box to be more like GeoTransform
Basically this is code copying. Unfortunately GeoReferenceBase can't
called GeoTransform for this calculation. Nor is the opposite correct.

We want to mimic GeoTransform just because we spent so much time
proving that it worked. (There is a special test generator for
image2qtree hidden in VW.) Unfortunately this comes at the cost of
being slower than Kyle's original transform. I think this is okay
sense this code is in GeoReferenceBase and that means it is generic,
thus we must be nearly exhaustive.