stereo: rewrite: Wrote optimized parabola subpixel correlator
This is to allow use of the new prefilters and to do an additional
performance increase. The performance is through the use of
prerasterization and using the fastbox sum instead of querying all
pixels by hand.

This gives a 2x to 3x speed improvement. Your mileage will vary based
on thread interactions as it seems now the system can't serve the image
fast enough.