stereo: rewrite: Stop bad malloc on correlation of too small a tile
The math for figuring out how many levels to processes was coming up
with a negative when the tile be rendered is too small (think around
or smaller than kernel size). That number casted to a size_t causes a
request for a really large array and a bad alloc.

There was also a problem where it was attempting to render a zone that
was outside the disparity image.

The pyramid data was too cropped to support cross consistency
check. Now it should refer back to the original image plus edge
extension.