Added NullProgressCallback, the new Dummy Instance
The cool thing about this Callback is that it does nothing and doesn't
use any mutexes. This is an important speed improvement for VW as a
lot of our functions like for_each_pixel take a ProgressCallback. When
the user gives that a TerminalProgressCallback that adds a mutex deep
inside a for loop. Yet if they don't provide anything, the default
constructor provides a DummyInstance that unfortunately was previously
mutexed. The old dummy instance still updated a progress that was
never queried.

This commit make the dummy instance request return a NullProgress
callback that is not mutex'd in anyway. This is a big deal because
this removes a lot of mutexs from a lot of loops.