Fix snapshot issue found in CTX Mosaic
The problem was that we stopped on the first opaque image that files
the entire tile. However inorder to get newest TID on top, we insert
low TIDs first and then high TIDs last. This meant that if a low TID
filled an entire tile, the newer TIDs would never be written.

This commit fixes the above problem and also keeps the opaque check
optimization in place. We know insert from top down and do a composite
every time a new tile is added. Then that composite is checked to see
if it is opaque. This is good as it means we might stop faster than
the previous code.

There's also a small improvement of never calling composite in the
event that there is only one input tile.