plate: Enforce use of return in plate tools
Using exit(0) or exit(1) is generally a bad idea as it doesn't call
destructors for local objects. In the case of exit(1), I replaced
those with vw_throw() so that users can potentially handle things
their own way.

Exit(0) were just changed out with return 0. This is so platefiles
have a chance to flush buffers and the like.

I also made sure all executables end with a defined return value.