plate: Make platecopy's full copy not overflow blob
When users do a full copy, platecopy now works out about how many
tiles 2GB is and then processes only in that many chunks. This way we
don't overflow a blob and the output platefile is still valid.

Also removed a warning of comparison between signed and unsigned
integers. The solution was just to store all levels as uint32.