fileio: Fix png comment loading
I also added a method to allow comment writing to PNG resources. I
then added a test to check for a complete cycle of writing a PNG
comment and then reading it all back in.

There's a little bit of a gimmick to it. You must write your comments
before you write the image data. I don't know enough about PNG to
truly understand this.