image: remove cstride from memory accessors
All instantiations had cstride set to one. With that option removed,
the compiler can simply increment instead of adding two numbers.

This provides a modest 40% improvement on gcc-4.2 for OSX for in-place
addition of two images. I'm assuming this benefits all other
algorithms as well. It's about a 300% improvement for gcc-mp-4.6 which
was previously making a huge mistake before that made it incredibly
slower than the apple supplied compiler.