plate: blobstore: stop possiblity of writing more than 4GB
This has blob store check on write_updates if its blob is going to
exceed 4 GB. If it is, it opens up a new blob for writing.

The reason we stop at 4 GB is because Blob.* uses 32bit offset
pointers. We could change that to be 64bit, however this way we're
taking control of when to start new blobs. As opposed to ... waiting
for a blob to exceed however many exobytes 64 bits is.