fileio: Fix DstMIROpenEXR::data
I was trying to return a pointer to std::ostringstream::str() which is
a temporary type. Mac 10.7 was cool with it, the rest of the world was
sadly not. :(

Oh well, I'm managing memory myself this time. I don't know how to get
a pointer to the inner buffer of an ostringstream. It doesn't seem
like the devs wanted you to or maybe it's doing something complex
internally.