cartography: Rework point image manipulation
The previous version of this file was inconsistent about if it was
using lla or llr elements. The use of radius was also extremely wrong
as that is an assumption for spherical datums. I've reworked the file
so that it is consistent about what is geodetic, cartesian,
projection, or point. This also allows me to correctly support earth
with point2dem in ASP.