Stereo: Fix mutex jam in SubpixelView
get_disparity_range pokes an image to death for calculating range
stats. We feed it a ImageViewRef and thus get_disparity_range hit a
mutex everytime it wanted a pixel because it hit direct cache. The
correct solution is to immediately render our desired chunks of memory
into the heap. Our cache is slow and does handle lots of theads poking
it to death.