stereo: Fix disparity masking in Pyramid correlation
There are primarily 2 things.

1.) Using edge_mask one the disparity masks limits the size of the
valid data that could be correlated. It's over constraining and
deletes wedges of data. I think the original author meant only to
remove data that was near the boundary of an image tile and not the
actual image itself. The commit replaces edge_mask with just a half
kernel masking from the image tile.

2.) Wrote a new subsample mask by 2. This one is a little cleaner with
memory access. It also allows the output to create a valid pixel if at
least 2 of the input are good. Previously we required that all input
pixels be good in order to make the output valid. That had the side
effect of making the mask slowly shrink as we continually subsampled.