stereo: Replace edge masking with average color in nodata.
It appears the edge masking was required to get good results along an
edge of an image where it meets the black of nodata. This commit still
removes the edge masking but replaces the nodata portion with a mean
color. This produces the same high quality of edge masking but still
allows us to process those final slivers of data. The slivers I'm
speaking of is a product of a map projected image and a tile boundary,
where the cut is in such a way that the tile is only fille maybe 30%
with real data.