Image: Update test and remove reference passing by ApplyMask
Reference passing by ApplyMask was breaking only on my 10.7 Macbook. I
believe it's the wrong thing to do because it means we're passing a
reference to type that only exists in the ApplyMask functor. When we
chain operations, that functor might die before the final output is
finished.

I'm not sure if that's the right interpretation. However I've written
a test to check for it and this commit also fixes it. I've also
updated the tests to be typed tests so we test all pixel types that
are interesting.