gui: Change hillshade and offset controls
Offset is now applied after gain. This makes the effect of offset
proportional to gain. It makes life a lot easier when viewing
disparities.

Hillshade is now a toggle through mulitple 'gains'. Since we don't
read georeference information, we can't pull out the best
normalization value. The user can instead just hit hillshade several
times until they hit a value that show just enough detail to not
saturate. It changes on a squared value so it should be pretty
general.