gui: image tile generator convert input to float instead of GL
The OpenGL loading of textures was going to do this anyway. The only
problem with OpenGL implementation is that they clamp the values. So
when loading int images, I would loose every thing outside of 0 and 1.

This works around that problem by just having us convert to float and
openGL can just copy instead of clamping.