Image: Subsample rasterize in chunks
When subsample is greatly reducing an image, it must prerasterize a
large chunk of the input to create its small output. This has a
tendency to blow our memory requirements sky high. Now subsample will
rasterize in terms of tiles on the input so that our memory usage is
reasonable.