Image: Move subsample's chunk rasterization to preraster call
SubsampleView's chunk raster was only being called when subsample view
was the last image operation. In this commit I move it to the
preraster call. The major downside of this is that I now have a buffer
image view for every tile rastered. However I still have the chunk
processing so that we are not having DiskImageView or other operations
raster extremely large buffers. It's not a perfect solution but I
believe that it is still an improvement.