image2qtree: Have projection NONE for all modes
Projection NONE is no longer used as a place holder for georeferenced
images. That job is now for projection DEFAULT.

Projection NONE is for when the user doesn't really want a
projection. The just want a quadtree out in some format equal to
MODE. This is useful for crazy people who want to implement things
like a tie point picker.