SubpixelView: Relax restriction on images having same size
In BayesEMSubpixelView it was required that left and right images have the same size, unlike in ParabolaSubpixelView. I looked carefully at the code and this assumption seems unjustified, as when the images are cropped to get a patch they are edge-extended if they go out of bounds. A check using valgrind confirmed that there are no memory violations in case the images have different sizes. Removing this restriction