Cache: Fix dead locking problem.
First ... we made the mutex order go:

Line -> Cache -> And then weakly attempt other lines in the event of invalidating.

We also force the user to release handle manually

This is required because all our access patterns are using a pointer
just outside of a mutex. Thus, some other thread can show up and
delete the data behind our pointer.