Remove std.find and use faster boost::starts_with
Also used some other boost features to reduce the amount of code since I
was in the area. I walked in this part of the code since cppcheck
complained about std::find and preferred std::compare for speed. Well
since we use boost, boost::starts_with is a little easier to read.